Conflicts:
	bin/main/GA.class
	src/main/GA.java
	src/plots/islands/MigrationRate.java

+++ src/representations/Adjacency.java
@@ -121,7 +121,7 @@ public class Adjacency extends Chromosome {
 	public Adjacency clone() {
 		List<Integer> clonedList = new ArrayList<Integer>();
 		for (int i = 0; i < getPath().size(); i++) {
-			clonedList.add(getPath().get(i));
+			clonedList.add(getPath().get(i).intValue());
 		}
 		return new Adjacency(problem, clonedList);
 	}
